move pubsub/ under internal/ until it stabilizes
