+++ internal/pubsub/acks_test.go
@@ -22,8 +22,8 @@ import (
 	"sync"
 	"testing"
 
-	"github.com/google/go-cloud/pubsub"
-	"github.com/google/go-cloud/pubsub/driver"
+	"github.com/google/go-cloud/internal/pubsub"
+	"github.com/google/go-cloud/internal/pubsub/driver"
 )
 
 type ackingDriverSub struct {
