+++ internal/pubsub/pubsub_test.go
@@ -20,8 +20,8 @@ import (
 	"sync"
 	"testing"
 
-	"github.com/google/go-cloud/pubsub"
-	"github.com/google/go-cloud/pubsub/driver"
+	"github.com/google/go-cloud/internal/pubsub"
+	"github.com/google/go-cloud/internal/pubsub/driver"
 )
 
 type driverTopic struct {
